+++ src/controllers/LiteTaxController.php
@@ -11,6 +11,7 @@ use Craft;
 use craft\commerce\models\LiteTaxSettings;
 use craft\commerce\Plugin;
 use craft\errors\WrongEditionException;
+use craft\helpers\Localization;
 use craft\i18n\Locale;
 use yii\base\Exception;
 use yii\web\Response;
@@ -67,6 +68,8 @@ class LiteTaxController extends BaseStoreSettingsController
 
         $percentSign = Craft::$app->getLocale()->getNumberSymbol(Locale::SYMBOL_PERCENT);
         $rate = Craft::$app->getRequest()->getBodyParam('taxRate');
+        $rate = Localization::normalizeNumber($rate);
+
         if (strpos($rate, $percentSign) || $rate >= 1) {
             $settings->taxRate = (float)$rate / 100;
         } else {