+++ src/controllers/DonationsController.php
@@ -30,7 +30,7 @@ class DonationsController extends BaseStoreSettingsController
      */
     public function actionEdit(array $variables = []): Response
     {
-        $donation = Donation::find()->one();
+        $donation = Donation::find()->status(null)->one();
 
         if ($donation === null) {
             $donation = new Donation();
@@ -52,7 +52,7 @@ class DonationsController extends BaseStoreSettingsController
         $this->requirePostRequest();
 
         // Not using a service to save a donation yet. Always editing the only donation.
-        $donation = Donation::find()->one();
+        $donation = Donation::find()->status(null)->one();
 
         if ($donation === null) {
             $donation = new Donation();
@@ -60,6 +60,7 @@ class DonationsController extends BaseStoreSettingsController
 
         $donation->sku = Craft::$app->getRequest()->getBodyParam('sku');
         $donation->availableForPurchase = (bool)Craft::$app->getRequest()->getBodyParam('availableForPurchase');
+        $donation->enabled = (bool)Craft::$app->getRequest()->getBodyParam('enabled');
 
         if (!Craft::$app->getElements()->saveElement($donation)) {
             return $this->renderTemplate('commerce/store-settings/donation/_edit', compact('donation'));