+++ src/models/Settings.php
@@ -84,7 +84,8 @@ class Settings extends Model
     public bool $allowCheckoutWithoutPayment = false;
 
     /**
-     * @var bool Whether partial payment can be made from the front end. Gateway must also allow them.
+     * @var bool Whether [partial payment](https://craftcms.com/docs/commerce/3.x/making-payments.html#checkout-with-partial-payment)
+     * can be made from the front end when the gateway allows them.
      *
      * The default `false` does not allow partial payments on the front end.
      *
@@ -205,7 +206,7 @@ class Settings extends Model
     public string $gatewayPostRedirectTemplate = '';
 
     /**
-     * @var string|null Default URL to be loaded after using the [load cart controller action](loading-a-cart.md).
+     * @var string|null Default URL to be loaded after using the [load cart controller action](orders-carts.md#loading-a-cart).
      *
      * If `null` (default), Craft’s default [`siteUrl`](config3:siteUrl) will be used.
      *