+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 - Fixed a bug where gateways’ `supportsPartialPayment()` methods weren’t being respected.
 - Fixed an error that could occur when saving a discount. ([#2660](https://github.com/craftcms/commerce/issues/2660))
 - Fixed a bug where partial payment errors weren’t getting returned correctly for Ajax requests to `commerce/payments/pay`.
+- Fixed an error that could occur when trying to refund an order. ([#2642](https://github.com/craftcms/commerce/pull/2642))
 
 ### Security
 - Fixed XSS vulnerabilities.