+++ CHANGELOG.md
@@ -2,6 +2,18 @@
 
 ## Unreleased
 
+### Added
+- Added the ability to enable/disable the Donation purchasable. ([#2374](https://github.com/craftcms/commerce/discussions/2374))
+- Added support for searching orders by line item description. ([#2658](https://github.com/craftcms/commerce/pull/2658))
+- Added `craft\commerce\elements\Order::isPaymentAmountPartial()`.
+
+### Fixed
+- Fixed a bug where gateways’ `supportsPartialPayment()` methods weren’t being respected.
+- Fixed an error that could occur when saving a discount. ([#2660](https://github.com/craftcms/commerce/issues/2660))
+- Fixed a bug where partial payment errors weren’t getting returned correctly for Ajax requests to `commerce/payments/pay`.
+- Fixed an error that could occur when trying to refund an order. ([#2642](https://github.com/craftcms/commerce/pull/2642))
+- Fixed a bug where tax rates weren’t properly supporting localized number formats.
+
 ### Security
 - Fixed XSS vulnerabilities.